H.Res. 233 Engrossed in House
Providing for consideration of the bill (H.R. 1620) to reauthorize the Violence Against Women Act of 1994, and for other purposes; providing for consideration of the bill (H.R. 6) to authorize the cancellation of removal and adjustment of status of certain aliens, and for other purposes; providing for consideration of the bill (H.R. 1603) to amend the Immigration and Nationality Act to provide for terms and conditions for nonimmigrant workers performing agricultural labor or services, and for other purposes; providing for consideration of the bill (H.R. 1868) to prevent across-the-board direct spending cuts, and for other purposes; providing for consideration of the joint resolution (H.J. Res. 17) removing the deadline for the ratification of the equal rights amendment; and for other purposes.

Version:  Engrossed in House (eh)
Date:     2021-03-16
Congress: 117

Source: https://www.govinfo.gov/content/pkg/BILLS-117hres233eh/xml/BILLS-117hres233eh.xml

Sponsored-By: T000474