H.Res. 403 Reported in House
Providing for consideration of the bill (H.R. 1629) to amend the Federal Food, Drug, and Cosmetic Act with respect to limitations on exclusive approval or licensure of orphan drugs, and for other purposes; providing for consideration of the resolution (H. Res. 275) condemning the horrific shootings in Atlanta, Georgia, on March 16, 2021, and reaffirming the House of Representative's commitment to combating hate, bigotry, and violence against the Asian-American and Pacific Islander community; and for other purposes.

Version:  Reported in House (rh)
Date:     2021-05-17
Congress: 117

Source: https://www.govinfo.gov/content/pkg/BILLS-117hres403rh/xml/BILLS-117hres403rh.xml

Sponsored-By: S001205