H.Res. 504 Engrossed in House
Providing for consideration of the bill (H.R. 2662) to amend the Inspector General Act of 1978, and for other purposes; providing for consideration of the bill (H.R. 3005) to direct the Joint Committee on the Library to replace the bust of Roger Brooke Taney in the Old Supreme Court Chamber of the United States Capitol with a bust of Thurgood Marshall to be obtained by the Joint Committee on the Library and to remove certain statues from areas of the United States Capitol which are accessible to the public, to remove all statues of individuals who voluntarily served the Confederate States of America from display in the United States Capitol, and for other purposes; providing for consideration of the bill (H.R. 3684) to authorize funds for Federal-aid highways, highway safety programs, and transit programs, and for other purposes; providing for consideration of the resolution (H. Res. 503) Establishing the Select Committee to Investigate the January 6th Attack on the United States Capitol; and for other purposes.

Version:  Engrossed in House (eh)
Date:     2021-06-29
Congress: 117

Source: https://www.govinfo.gov/content/pkg/BILLS-117hres504eh/xml/BILLS-117hres504eh.xml

Sponsored-By: S001205
Roll-Call: House 117-1-189 2021-06-29 Passed 214-195
Roll-Call: House 117-1-190 2021-06-29 Passed 218-197