H.Res. 601 Engrossed in House
Providing for consideration of the bill (H.R. 4) to amend the Voting Rights Act of 1965 to revise the criteria for determining which States and political subdivisions are subject to section 4 of the Act, and for other purposes; providing for consideration of the Senate amendment to the bill (H.R. 3684) to authorize funds for Federal-aid highways, highway safety programs, and transit programs, and for other purposes; and providing for the adoption of the concurrent resolution (S. Con. Res. 14) setting forth the congressional budget for the United States Government for fiscal year 2022 and setting forth the appropriate budgetary levels for fiscal years 2023 through 2031; and for other purposes.

Version:  Engrossed in House (eh)
Date:     2021-08-24
Congress: 117

Source: https://www.govinfo.gov/content/pkg/BILLS-117hres601eh/xml/BILLS-117hres601eh.xml

Sponsored-By: N000191
Roll-Call: House 117-1-257 2021-08-24 Passed 220-212
Roll-Call: House 117-1-258 2021-08-24 Passed 220-212