H.Res. 973 Reported in House
Providing for consideration of Senate amendment to the bill (H.R. 2471) to measure the progress of post-disaster recovery and efforts to address corruption, governance, rule of law, and media freedoms in Haiti; providing for consideration of the joint resolution (H.J. Res. 75) making further continuing appropriations for the fiscal year ending September 30, 2022, and for other purposes; providing for consideration of the bill (H.R. 6968) to prohibit the importation of energy products of the Russian Federation, and for other purposes; and providing for consideration of the bill (H.R. 7007) making emergency supplemental appropriations for coronavirus response and relief for the fiscal year ending September 30, 2022, and for other purposes.

Version:  Reported in House (rh)
Date:     2022-03-09
Congress: 117

Source: https://www.govinfo.gov/content/pkg/BILLS-117hres973rh/xml/BILLS-117hres973rh.xml

Sponsored-By: M000312