H.Res. 972 Reported in House
Providing for consideration of Senate amendment to the bill (H.R. 2471) to measure the progress of post-disaster recovery and efforts to address corruption, governance, rule of law, and media freedoms in Haiti; providing for consideration of the joint resolution (H.J. Res. 75) making further continuing appropriations for the fiscal year ending September 30, 2022, and for other purpose; and providing for consideration of the bill (H.R. 6968) to prohibit the importation of energy products of the Russian Federation, and for other purposes.

Version:  Reported in House (rh)
Date:     2022-03-09
Congress: 117

Source: https://www.govinfo.gov/content/pkg/BILLS-117hres972rh/xml/BILLS-117hres972rh.xml

Sponsored-By: M000312