S. 5085 Introduced in Senate
A bill to prohibit the government of the District of Columbia from using Federal funds to allow individuals who are not citizens of the United States to vote in any election. and for other purposes.

Version:  Introduced in Senate (is)
Date:     2022-11-14
Congress: 117

Source: https://www.govinfo.gov/content/pkg/BILLS-117s5085is/xml/BILLS-117s5085is.xml

Sponsored-By: C001098