H.Res. 1499 Reported in House
Providing for consideration of the bill (H.R. 3372) to authorize implementation grants to community-based nonprofits to operate one-stop reentry centers; providing for consideration of the bill (H.R. 6878) to address the health needs of incarcerated women related to pregnancy and childbirth,and for other purposes; providing for consideration of the bill (H.R. 8876) to reauthorize the Maternal, Infant, and Early Childhood Home Visiting program, and for other purposes; providing for consideration of the joint resolution (H.J. Res. 100) to provide for a resolution with respect to the unresolved disputes between certain railroads represented by the National Carriers' Conference Committee of the National Railway Labor Conference and certain of their employees; and for other purposes.

Version:  Reported in House (rh)
Date:     2022-11-30
Congress: 117

Source: https://www.govinfo.gov/content/pkg/BILLS-117hres1499rh/xml/BILLS-117hres1499rh.xml

Sponsored-By: M000312